chore(webpack): extract commons
